Remove nested blueprints from sliders
